Add check for mount permissions
It's possible that there are some issues while mounting paths when a container
runs in its own user namespace. It happens because the remapped id on the host
could not have the permissions to read or transverse a path.

This commit adds and explicit check for that to print a clear error message.